Break out some logic into utils
